+++ easygui/boxes/state.py
@@ -1,0 +1,23 @@
+"""
+
+.. moduleauthor:: Stephen Raymond Ferg and Robert Lugg (active)
+.. default-domain:: py
+.. highlight:: python
+
+Version |release|
+"""
+
+# Starting and global variables
+
+rootWindowPosition = "+300+200"
+
+PROPORTIONAL_FONT_FAMILY = ("MS", "Sans", "Serif")
+MONOSPACE_FONT_FAMILY = ("Courier")
+
+PROPORTIONAL_FONT_SIZE = 10
+# a little smaller, because it is more legible at a smaller size
+MONOSPACE_FONT_SIZE = 9
+TEXT_ENTRY_FONT_SIZE = 12  # a little larger makes it easier to see
+
+
+STANDARD_SELECTION_EVENTS = ["Return", "Button-1", "space"]
